Refactor date filter into a reusable component
