+++ src/apps/media/src/app/components/MediaList.tsx
@@ -24,7 +24,6 @@ import {
   fileExtension,
   fileTypeToColor,
   getExtensions,
-  getDateFilterFn,
 } from "../utils/fileUtils";
 import { AppState } from "../../../../../shell/store/types";
 import {
