+++ src/apps/media/src/app/views/AllMedia.tsx
@@ -17,11 +17,10 @@ import Controls from "../components/Controls";
 import { useMemo, useState } from "react";
 import { AppState } from "../../../../../shell/store/types";
 import {
-  fileExtension,
-  getExtensions,
   getDateFilterFn,
   getDateFilter,
-} from "../utils/fileUtils";
+} from "../../../../../utility/dateUtils";
+import { fileExtension, getExtensions } from "../utils/fileUtils";
 import { Filetype } from "../../../../../shell/store/media-revamp";
 import { useParams as useSearchParams } from "../../../../../shell/hooks/useParams";
 import { State } from "../../../../../shell/store/media-revamp";
