+++ src/shell/components/DateRangeFilter/DateFilter.tsx
@@ -12,12 +12,12 @@ import CloseRounded from "@mui/icons-material/CloseRounded";
 import CheckIcon from "@mui/icons-material/Check";
 import Divider from "@mui/material/Divider";
 
-import { AppState } from "../../../../../../shell/store/types";
-import { DateRange } from "../../../../../../shell/store/media-revamp";
-import { getDateFilter } from "../../utils/fileUtils";
+import { AppState } from "../../store/types";
+import { DateRange } from "../../store/media-revamp";
+import { getDateFilter } from "../../../utility/dateUtils";
 import moment from "moment-timezone";
-import { DateFilterModal } from "../DateFilterModal";
-import { useParams } from "../../../../../../shell/hooks/useParams";
+import { DateFilterModal } from "./DateFilterModal";
+import { useParams } from "../../hooks/useParams";
 
 type Modal = "on" | "before" | "after" | null;
 export const DateRangeFilter: FC = () => {
